* feat(123-windows-port): add some auto pointer

* fix(PIM-1234): add lambda inactive flag

* fix(PIM-1234): merge main

* Merge branch 'main' of https://github.com/jensvogt/awsmock

* Merge branch 'main' into feat(123-windows-port)

* feat(123-windows-port): remove COREs for frontend

* Merge remote-tracking branch 'origin/main'

* Merge branch 'main' into feat(123-windows-port)

* feat(123-windows-port): remove COREs for frontend

* feat(123-windows-port): remove COREs for frontend

* feat(123-windows-port): first try to port it to Win32

* Merge branch 'main' into feat(123-windows-port)

+++ src/dto/include/awsmock/dto/kms/model/Key.h
@@ -108,7 +108,7 @@ namespace AwsMock::Dto::KMS {
         /**
          * Creation date
          */
-        long creationDate = system_clock::to_time_t(system_clock::now());
+        long creationDate = static_cast<long>(system_clock::to_time_t(system_clock::now()));
 
         /**
          * Deletion date
